app/backend/requirements.txt
numpy
fastapi
uvicorn
nibabel
python-dotenv
pillow
ninja
scipy
pandas
python-multipart
pytest
httpx